libimcv: Cast chunk length to int when printing as string

+++ src/libimcv/plugins/imv_swid/imv_swid_state.c
@@ -301,7 +301,8 @@ METHOD(imv_swid_state_t, set_swid_inventory, void,
 		tag_creator = tag_id->get_tag_creator(tag_id);
 		sw_id = tag_id->get_unique_sw_id(tag_id, NULL);
 		snprintf(software_id, BUF_LEN, "%.*s__%.*s",
-				 tag_creator.len, tag_creator.ptr, sw_id.len, sw_id.ptr);
+				 (int)tag_creator.len, tag_creator.ptr,
+				 (int)sw_id.len, sw_id.ptr);
 		DBG3(DBG_IMV, "  %s", software_id);
 
 		/* Add software ID to JSON array */
